--- src/features/chefs/index.tsx
@@ -1,9 +1,0 @@
-//component
-export * from "./components/chef-list/vertical-chef-list-item";
-export * from "./components/chef-list/vertical-chef-list";
-
-// lib
-export * from "./lib/actions";
-
-//type
-export * from "./types";
